+++ NEWS.md
@@ -11,6 +11,7 @@
 * `ir_correct_atmosphere()` has a new parameter `return_contribution` which allows to return the contribution of `ref` to each spectrum in `x`.
 * `ir_normalize()` has a new `method` to normalize spectra: With `method = "area_absolute"`, all intensity values will be divided by the sum of the absolute intensity values at all wavenumber values of the spectrum.
 * `ir_bc_rubberband()` has an additional parameter to allow imputation of the first and last value in a spectra to avoid artifacts which may happen during baseline correction with the rubberband algorithm.
+* All functions now fully support empty spectra.
 
 # ir 0.3.0